refactor: moved logic from data_ui_generator to respective child classes
