+++ splunk_add_on_ucc_framework/generators/xml_files/create_inputs_xml.py
@@ -15,7 +15,7 @@
 #
 from splunk_add_on_ucc_framework.generators.file_generator import FileGenerator
 from typing import Dict
-from splunk_add_on_ucc_framework import data_ui_generator
+from lxml import etree as ET
 
 
 class InputsXml(FileGenerator):
@@ -24,9 +24,26 @@ class InputsXml(FileGenerator):
         " in `default/data/ui/views/inputs.xml` folder"
     )
 
+    def generate_views_inputs_xml(self, addon_name: str) -> str:
+        """
+        Generates `default/data/ui/views/inputs.xml` file.
+        """
+        view = ET.Element(
+            "view",
+            attrib={
+                "template": f"{addon_name}:/templates/base.html",
+                "type": "html",
+                "isDashboard": "False",
+            },
+        )
+        label = ET.SubElement(view, "label")
+        label.text = "Inputs"
+        view_as_string = ET.tostring(view, encoding="unicode", pretty_print=True)
+        return view_as_string
+
     def _set_attributes(self) -> None:
         if self._global_config.has_inputs():
-            self.inputs_xml_content = data_ui_generator.generate_views_inputs_xml(
+            self.inputs_xml_content = self.generate_views_inputs_xml(
                 self._addon_name,
             )
